requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
nibabel>=3.2.0
matplotlib>=3.4.0
Pillow>=8.3.0
tqdm>=4.61.0
requests>=2.25.0
scikit-learn>=0.24.0
lpips>=0.1.4
pytorch-msssim>=0.2.1
torchmetrics>=0.11.0